Bedroom curtains now work with new iteminteraction script
