+++ Assets/Animations/Curtains.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ea1d3a2afe9dd343bc3f82d980f4dd8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
