+++ Assets/Animations/Yellow Curtains.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a109e63685046a49998ca3c1abd635b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
