+++ Assets/Animations/Yellow Curtains/CurtainsActive.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 08004e9106216494a90a67f1041da9cf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
