+++ Assets/Animations/YellowCurtains.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d2def842a3b75f439bd6a6c6ddd927c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
