+++ Assets/EnvironmentalItem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fcf5c751d42916a48b832ebf1643315e
+guid: 9432028c78ffe904abe8fd7a565a5d44
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
